Add data + queries

+++ schema.sql
@@ -1,7 +1,7 @@
 /* Database schema to keep the structure of entire database. */
 
 CREATE TABLE animals (
-  id int,    
+  id BIGSERIAL,    
   name varchar(100),
   date_of_birth date,
   escape_attempts int,
